validations and checks

+++ controllers/groupController.js
@@ -2,7 +2,7 @@
 const groupModel = require('../models/groupModel.js');
 const discModel = require('../models/discussionModel.js');
 const userModel = require('../models/userModel.js');
-//GET ALL GROUPS
+//GET ALL GROUPS INFO
 exports.GetInfo = async (req, res) => {
   try {
     const group = await groupModel.find();
@@ -22,6 +22,14 @@ exports.GetInfo = async (req, res) => {
 //GET MEMBERS
 exports.GetMembers = async (req, res) => {
   try {
+    if ((await groupModel.findById(req.params.id)) === null) {
+      res.status(400).send({
+        status: 'error',
+        message: 'Couldnt find group',
+      });
+      return;
+    }
+
     const members = await groupModel
       .findById(req.params.id)
       .select({ _id: 0 })
@@ -43,6 +51,13 @@ exports.GetMembers = async (req, res) => {
 
 exports.GetPhotoPool = async (req, res) => {
   try {
+    if ((await groupModel.findById(req.params.id)) === null) {
+      res.status(400).send({
+        status: 'error',
+        message: 'Couldnt find group',
+      });
+      return;
+    }
     const photos = await groupModel
       .findById(req.params.id)
       .select({ _id: 0 })
@@ -63,6 +78,13 @@ exports.GetPhotoPool = async (req, res) => {
 //GET ALL DISCUSSIONS IN A GROUP
 exports.GetAllDiscussions = async (req, res) => {
   try {
+    if ((await groupModel.findById(req.params.id)) === null) {
+      res.status(400).send({
+        status: 'error',
+        message: 'Couldnt find group',
+      });
+      return;
+    }
     const discussions = await groupModel
       .findById(req.params.id)
       .select({ _id: 0 })
@@ -83,6 +105,13 @@ exports.GetAllDiscussions = async (req, res) => {
 //DELETE A DISCUSSION BY ID
 exports.DeleteDiscussion = async (req, res) => {
   try {
+    if ((await discModel.findById(req.params.id)) === null) {
+      res.status(400).send({
+        status: 'error',
+        message: 'discussion doesnt exist',
+      });
+      return;
+    }
     await discModel.findByIdAndDelete(req.params.id);
     res.status(200).send({
       status: 'success',
@@ -99,6 +128,13 @@ exports.DeleteDiscussion = async (req, res) => {
 //GET A DISCUSSION BY ID
 exports.getDiscussion = async (req, res) => {
   try {
+    if ((await discModel.findById(req.params.id)) === null) {
+      res.status(400).send({
+        status: 'error',
+        message: 'discussion doesnt exist',
+      });
+      return;
+    }
     const discussion = await discModel
       .findById(req.params.id)
       .select({ _id: 0 });
@@ -118,7 +154,26 @@ exports.getDiscussion = async (req, res) => {
 //CREATE DISCUSSION
 exports.createDiscussion = async (req, res) => {
   try {
-    const newDiscussion = await discModel.create(req.body);
+    const newDiscussion = await discModel.create(req.body); //create new discussion instance
+    //if newDiscussion id is duplicated
+    if (await discModel.findById(newDiscussion)) {
+      res.status(400).send({
+        status: 'error',
+        message: 'ID is duplicated',
+      });
+      return;
+    }
+
+    //if group doesnt exist,end
+    if ((await groupModel.findById(req.params.id)) === null) {
+      res.status(400).send({
+        status: 'error',
+        message: 'group doesnt exist',
+      });
+      return;
+    }
+
+    //else add new discusion to the group
     await groupModel.findByIdAndUpdate(
       req.params.id,
       {
@@ -143,6 +198,15 @@ exports.createDiscussion = async (req, res) => {
 //EDIT A DISCUSSION
 exports.EditDiscussion = async (req, res) => {
   try {
+    //if discussion doesnt exist, end
+    if ((await discModel.findById(req.params.id)) === null) {
+      res.status(400).send({
+        status: 'error',
+        message: 'discussion doesnt exist',
+      });
+      return;
+    }
+    //else update
     const newDisc = await discModel.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
       runValidators: true,
